requirements.txt
flask
opencv-python-headless
paho-mqtt
numpy
ntplib
oracledb
